requirements.txt
altair==5.5.0
asgiref==3.11.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.12.0
blinker==1.9.0
boto3==1.42.19
botocore==1.42.19
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
narwhals==1.39.0
nest-asyncio==1.6.0
npm==0.1.1
numpy==2.2.5
optional-django==0.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytokens==0.3.0
pytz==2025.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.16.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sqlparse==0.5.5
stack-data==0.6.3
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
whitenoise==6.11.0
zipp==3.20.0
